Pushed the function defintions into their own cpp file
