+++ Gems/Prefab/PrefabDependencyViewer/Code/Source/DirectedGraph.h
@@ -39,62 +39,12 @@ namespace PrefabDependencyViewer::Utils
     {
     public:
         DirectedTree() = default;
-
-        void SetRoot(AZStd::shared_ptr<Node> root) {
-            if (m_root) {
-                AZ_Assert(false, "Prefab Dependency Viewer - Memory leak in the graph because the root was already set.");
-            }
-
-            m_root = root;
-        }
-
-        AZStd::shared_ptr<Node> GetRoot() const
-        {
-            return m_root;
-        }
-
         ~DirectedTree() = default;
 
-        AZStd::tuple<AZStd::vector<int>, int> countNodesAtEachLevel() const
-        {
-            /** Directed Graph can't have cycles because of the
-            non-circular nature of the Prefab. */
-            int widestLevelSize = 0;
-            AZStd::vector<int> count;
+        void SetRoot(AZStd::shared_ptr<Node> root);
+        AZStd::shared_ptr<Node> GetRoot() const;
+        AZStd::tuple<AZStd::vector<int>, int> countNodesAtEachLevel() const;
 
-            using pair = AZStd::pair<int, Node*>;
-            AZStd::queue<pair> queue;
-            queue.emplace(0, m_root.get());
-
-            while (!queue.empty())
-            {
-                auto [level, currNode] = queue.front();
-                queue.pop();
-
-                if (count.size() <= level)
-                {
-                    // Started a new level so check if the previous level
-                    // was any bigger then the widest level size seen so far.
-                    int prevLevelCount = level != 0 ? count[level - 1] : 0;
-                    widestLevelSize = AZStd::max(widestLevelSize, prevLevelCount);
-                    count.push_back(1);
-                }
-                else
-                {
-                    ++count[level];
-                }
-
-                auto children = currNode->GetChildren();
-                for (const NodePtr& node : children)
-                {
-                    queue.emplace(level + 1, node.get());
-                }
-            }
-
-            // Check if the last level was the widest.
-            widestLevelSize = AZStd::max(widestLevelSize, count[count.size() - 1]);
-            return AZStd::make_tuple(count, widestLevelSize);
-        }
     private:
         NodePtr m_root = nullptr;
     };
